Refactor Transaction module
